Move the border from header to whole intro
Move the border from being along the text element to being along the entire
container div such that the border will span the entire content div and
underline the avatar in the horizontal desktop layout.

Readjust the layout of the avatar and header by center-aligning them both in the
vertical and horizontal axes using flexbox.